Revert "Revert "ResourceParser""
This reverts commit 1338f1e5db38cace26949a87047ebfdb5e1792a8.
